Expose shapeless :: and HNil through Formality singleton for patterns.
We expose :: as :+: instead, which matches the name of the operator in
lift-common's HList implementation. This allows pattern matching the
HList in the callback handlers for form submission. While the
FormUnapplier may often be used to get a case class out of the success
case, you'll still need the operators/HNil to pattern match in failure
and generic submission handlers unless you want to write a version of
your case class that is the same but with Boxed parameters.