Make value conversion/serialization be via instance methods.
Select boxes will require us to do these by manipulating
instance variables, rather than by using external functions.
We also eliminate the ValueSerializerType concept; this was
used so that there could be a value serializer that serialized
a superclass of the FieldValueType instead of only being
able to use one that serialized the FieldValueType directly.

We may reintroduce it later, but for now dropping the
concept makes it easier to move value serialization into
an instance method.