Added support for radio buttons in selectField.
Right now this only covers selectField when invoked with
SelectableOptions, but it's only specs and a small amount of
glue code missing for tuple lists and regular lists.

Radio buttons are created by passing an
asRadioButtons = true parameter to selectField. That
parameter repeates the element selected by the passed
selector, binding a label and radio button within once for
each of the provided options.