Don't share field value TransientRequestVar between fields.
Before we were using a single private object. Though it was declared
under the FieldValueBase class, this didn't make the session field name
different per field. To fix this, we now put a new instance of a
TransientRequestVar per field. This instance is assigned a random name
so that it is tracked uniquely in the session.

This fixes an issue where having multiple fields of the same type would
result in all of them getting the same value.