Add FieldGroup and Formality.fieldGroup().
FieldGroup is used to group a set of fields that can then
return a single type. By default, they are simply an HList
inside the regular fields HList; however, you can use a
withConverter function to take the provided fields and convert
them to a more complex type. This type then makes its way
into the final HList instead of the nested HList.

This can be used, for example, to build case classes directly
from a group of fields, and then deal with the case class
instance directly in the success handler. FieldGroups are
themselves fields, as they implement the FieldHolderBase[T]
trait, with T being their converted type.

The converter returns a Box[T]; if it returns a Failure, then the
form conversion is considered failed and that failure can then
be passed on. Additionally, if any of the fields within fail their
validations, you get a failure for the group.

Also bump the example to 0.3.0-SNAPSHOT and use field
groups in the login screen.