Add field group unboxed conversion and "as".
as can be used to pass in a converter function that doesn’t
produce a Box[T], but just a naked T. It is automatically
wrapped in a Full. This is useful for case classes, where the
apply method does not return a Box.

Improve chat example’s signup snippet to use this.