Add boxed converter support to FieldGroup.
A boxed converter takes the boxed versions of the fields. It can be very useful
to implement a situation where the user can set one of two fields, and the
result is contingent on both fields. One example is a situation where something
can be either pasted into a text box or uploaded---the result may always be one
String (or, conversely, always be one array of bytes), but we have to choose
between two fields, one of which was potentially left empty. The current field
group converter, which only takes the unboxed parameter values, would simply
always fail in these cases.