Add specs for field gropus with too many fields for regular functions.
These are field groups consisting of >22 fields, which means we can only add
converters that are based in HLists. In this case, we don't support `as`,
`withConverter`, or `withBoxedConverter`, instead requiring
`withHlistConverter` and `withBoxedHlistConverter` to be invoked directly.

We do this because the types will be inferred for these correctly, while if we
rely on the macro-generated functions used above, they will not be inferred
correctly. This is more or less okay in cases where we're doing a regular
function, but having to manually declare an HList type is just nasty, so we
simply don't support those methods in that case.