Add a friendly error if we can't resolve the field group types.
This happens when a field group is declared in one place and the converter is
strapped onto it at a later time. I'm not actually sure *why* this happens, but
for now we'll go for a friendly error message over an opaque, difficult to
figure out message (that bit me, incidentally...).