Leave value attribute unchanged when no default field value is specified.
Until now, not specifying a default field value would clear any default value
that was specified in the markup. We now adhere to our principle of respecting
the markup in all possible cases by leaving the default value in the markup
untouched unless the caller explicitly sets a default value (which can be
blank for string values).

In cases where the caller wants to specify a blank default, and the markup has
a value specified, they can use a separate CSS transform to clear the value and
be explicit about that preference.