Rework how we represent HNil.
Scala 2.12 has some tweaked exhaustiveness checks on pattern matching that
catch a possibility we hadn't aniticipated. In particular, HNil was defined as
a final class, and we created a singleton instance of it to reference, but a
user could instantiate the final class themselves and send that in and wreak
havoc.

We redefine HNil as a final singleton object, and define a type HNil that is
equivalent to the singleton's type so we can reference it as both a value and
type called HNil. Other code remains unchanged, fortunately.

Unfortunately, this means we have to keep using our forkity fork of Lift's
HLists... Hopefully a later Lift minor version can include these tweaks, but
we'll need to maintain our fork until we no longer support versions of Lift
with the old implementation.