Auto-publish snapshots
We add a travis.sh build script that checks for this to be a non-PR build. The
build is only run on the master branch as a bonus. If all that is true AND we
are building a snapshot version, we descript the encrypted Sonatype credentials
and use them to publish the snapshot to Sonatype.